Add bank compression executable (#995)
* add spa compression; make mchirp calling, return more similar

* fix bug in FilterBank that ignores dtype

* add pycbc_compress_bank

* fixed copyright

* store correct values (#1)

* added precision to compress bank; removed fixed points option

* use end_idx for kmax

* add catch for discontinuity in time estimate

* have segment length computed on the fly

* moved compress_waveform to compress module

* compute segment length using log2

* fix bugs in seg length estimation and indexing

* fix bug in returning sample points; increase seg length for time estimate

+++ pycbc/waveform/bank.py
@@ -297,7 +297,7 @@ class FilterBank(TemplateBank):
 
         self.table[index].template_duration = template_duration        
 
-        htilde = htilde.astype(numpy.complex64)
+        htilde = htilde.astype(self.dtype)
         htilde.f_lower = self.f_lower
         htilde.end_frequency = f_end
         htilde.end_idx = int(htilde.end_frequency / htilde.delta_f)
